requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pytest>=8.0.0
pyyaml>=6.0.0
mlflow>=2.12.0
pillow>=10.0.0
nltk>=3.8.0
httpx>=0.27.0
